partially working with new components
