+++ client/scripts/models/index.ts
@@ -11,6 +11,7 @@ export { default as IWebWallet } from './IWebWallet';
 export { default as Identity } from './Identity';
 export { default as NodeInfo } from './NodeInfo';
 export { default as Notification } from './Notification';
+export { default as DashboardNotification } from './DashboardNotification';
 export { default as NotificationCategory } from './NotificationCategory';
 export { default as NotificationSubscription } from './NotificationSubscription';
 export { default as OffchainAttachment } from './OffchainAttachment';
