+++ client/styles/pages/user_dashboard.scss
@@ -117,10 +117,8 @@
     
 }
 
-.UserDashboardRow {
-    .comment-body {
-        .comment-body-bottom {
-            
-        }
-    }
+.Spinner {
+    margin-top: 20px;
+    display: flex;
+    justify-content: center;
 }
